+++ Services/Data/SongSerializer.cs
@@ -27,8 +27,8 @@ namespace Maestro.Services.Data
             {
                 Name = song.Name,
                 Artist = song.Artist,
+                Transcriber = song.Transcriber,
                 Instrument = song.Instrument.ToString(),
-                Bpm = song.Bpm,
                 Notes = song.Notes
             };
 
@@ -45,13 +45,14 @@ namespace Maestro.Services.Data
             {
                 Name = dto.Name,
                 Artist = dto.Artist,
+                Transcriber = dto.Transcriber,
                 Instrument = instrument,
-                Bpm = dto.Bpm
+                SkipOctaveReset = dto.SkipOctaveReset
             };
 
-            if (dto.Notes != null && dto.Bpm.HasValue)
+            if (dto.Notes != null)
             {
-                var commands = NoteParser.Parse(dto.Notes, dto.Bpm.Value);
+                var commands = NoteParser.Parse(dto.Notes);
                 song.Commands.AddRange(commands);
             }
 
@@ -73,13 +74,14 @@ namespace Maestro.Services.Data
                 {
                     Name = dto.Name,
                     Artist = dto.Artist,
+                    Transcriber = dto.Transcriber,
                     Instrument = instrument,
-                    Bpm = dto.Bpm
+                    SkipOctaveReset = dto.SkipOctaveReset
                 };
 
-                if (dto.Notes != null && dto.Bpm.HasValue)
+                if (dto.Notes != null)
                 {
-                    var commands = NoteParser.Parse(dto.Notes, dto.Bpm.Value);
+                    var commands = NoteParser.Parse(dto.Notes);
                     song.Commands.AddRange(commands);
                 }
 
@@ -97,14 +99,17 @@ namespace Maestro.Services.Data
             [JsonProperty("artist")]
             public string Artist { get; set; }
 
+            [JsonProperty("transcriber")]
+            public string Transcriber { get; set; }
+
             [JsonProperty("instrument")]
             public string Instrument { get; set; }
 
-            [JsonProperty("bpm")]
-            public int? Bpm { get; set; }
-
             [JsonProperty("notes")]
             public List<string> Notes { get; set; }
+
+            [JsonProperty("skipOctaveReset")]
+            public bool SkipOctaveReset { get; set; }
         }
     }
 }